colocate: ignore empty module names
When extracting module names with utils::getModuleName, any explicit or
implicit import of an 'index' file in a module's directory or one of its
parents resolves to an empty string. Since all JS strings begin with an
empty string, this is considered a matching module name for a used
fragment spread. Subsequent imports are ignored because a module name
matches. However, the falsiness of the empty string causes an error to
still be reported.

In pathological cases, an empty module name can be imported _before_ the
import that correctly matches a used fragment spread. This resulted in
a false-positive "must be colocated" error, despite the fact that the
required import is in fact present.

Ignore empty module names when looking for imports that satisfy a
fragment spread's colocation requirement.